+++ src/constants.js
@@ -1,0 +1,18 @@
+export const types = [
+  {
+    id: 0,
+    title: 'Any',
+  },
+  {
+    id: 1,
+    title: 'Administrative force majeure leave',
+  },
+  { id: 2, title: 'Administrative leave' },
+  { id: 3, title: 'Social leave' },
+  { id: 4, title: 'Sick leave (no documents)' },
+  { id: 5, title: 'Sick leave (with documents)' },
+  { id: 6, title: 'Study leave' },
+  { id: 7, title: 'Paid leave' },
+];
+
+export const states = ['Any', 'New', 'In progress', 'Approved', 'Rejected'];
